refactor: reorganize comon substrate zod parsers and shared types
- Move Api type and SbQueryError from fees.ts to dedicated modules
- Consolidate all base types from types/base.ts into chain/common/types.ts
- Create dedicated chain/common/errors.ts for error types
- Remove redundant sb_amount schema and Amount type
- Simplify sb_balance to use sb_bigint directly without nonnegative constraint
- Update governance cost field to use sb_bigint instead of sb_amount
- Fix imports across 8+ files for reorganized type locations
- Maintain backward compatibility through re-exports in types/index.ts

BREAKING: Types previously imported from types/base.ts must now be imported
from types/index.ts (re-exported) or directly from chain/common/types.ts
