+++ packages/torus-sdk-ts/src/chain/common/fees.ts
@@ -1,31 +1,11 @@
-import type { ApiPromise } from "@polkadot/api";
-import type {
-  ApiDecoration,
-  SignerOptions,
-  SubmittableExtrinsic,
-} from "@polkadot/api/types";
+import type { SignerOptions, SubmittableExtrinsic } from "@polkadot/api/types";
 import type { AccountId, Extrinsic } from "@polkadot/types/interfaces";
 import type { Result } from "@torus-network/torus-utils/result";
 import { makeErr, makeOk } from "@torus-network/torus-utils/result";
 import { tryAsync } from "@torus-network/torus-utils/try-catch";
 import { sb_balance } from "../../types/index.js";
-
-export type Api = ApiDecoration<"promise"> | ApiPromise;
-
-export * from "./storage-maps.js";
-
-// ==== Error ====
-
-export class SbQueryError extends Error {
-  constructor(message: string, options?: ErrorOptions) {
-    super(message, options);
-    this.name = "SbQueryError";
-  }
-
-  static from(error: Error): SbQueryError {
-    return new SbQueryError(error.message, { cause: error });
-  }
-}
+import { SbQueryError } from "./errors.js";
+import type { Api } from "./types.js";
 
 // ==== Extrinsic fees ====
