+++ packages/torus-sdk-ts/src/chain/governance/governance-storage.ts
@@ -8,7 +8,7 @@ import {
 import type { SS58Address } from "../../types/address.js";
 import { sb_address, sb_array, sb_id, sb_some } from "../../types/index.js";
 import { queryFreeBalance } from "../balances.js";
-import type { Api } from "../common/fees.js";
+import type { Api } from "../common/types.js";
 import { handleMapValues } from "../common/storage-maps.js";
 import type {
   AgentApplication,
